refactor(presentation): move narrow tab layout helpers to their own module
