+++ packages/sanity/src/presentation/PresentationNavigator.tsx
@@ -2,7 +2,7 @@ import {memo, useCallback, useMemo} from 'react'
 
 import {Panel} from './panels/Panel'
 import {PanelResizer} from './panels/PanelResizer'
-import {getPresentationPanelHtmlId} from './panels/PresentationNarrowTabBar'
+import {getPresentationPanelHtmlId} from './panels/presentationLayoutTab'
 import {type NavigatorOptions} from './types'
 import {useLocalState} from './useLocalState'
